add dilithium

+++ src/include/libfam/sign.h
@@ -47,19 +47,17 @@ i32 pqcrystals_dilithium2_ref_keypair(u8 *pk, u8 *sk, const u8 seed[32]);
 i32 pqcrystals_dilithium2_ref_signature(u8 *sig, u64 *siglen, const u8 *m,
 					u64 mlen, const u8 *ctx, u64 ctxlen,
 					const u8 *sk, Rng *rng);
-i32 pqcrystals_dilithium2_ref_verify(const uint8_t *sig, size_t siglen,
-				     const uint8_t *m, size_t mlen,
-				     const uint8_t *ctx, size_t ctxlen,
-				     const uint8_t *pk);
+i32 pqcrystals_dilithium2_ref_verify(const u8 *sig, u64 siglen, const u8 *m,
+				     u64 mlen, const u8 *ctx, u64 ctxlen,
+				     const u8 *pk);
 
 i32 pqcrystals_dilithium2_avx2_keypair(u8 *pk, u8 *sk, const u8 seed[32]);
 i32 pqcrystals_dilithium2_avx2_signature(u8 *sig, u64 *siglen, const u8 *m,
 					 u64 mlen, const u8 *ctx, u64 ctxlen,
 					 const u8 *sk, Rng *rng);
-i32 pqcrystals_dilithium2_avx2_verify(const uint8_t *sig, size_t siglen,
-				      const uint8_t *m, size_t mlen,
-				      const uint8_t *ctx, size_t ctxlen,
-				      const uint8_t *pk);
+i32 pqcrystals_dilithium2_avx2_verify(const u8 *sig, u64 siglen, const u8 *m,
+				      u64 mlen, const u8 *ctx, u64 ctxlen,
+				      const u8 *pk);
 
 typedef struct {
 	__attribute__((aligned(32))) u8 data[DILITHIUM_SECRETKEY_SIZE];
